Ton of fixes and cleanup
- Reuse generic approval view for request approval view.
- Delete redundant view controllers (i.e. when UIHostingController suffices).
- Unwrap closures that invoke a single no-arg function.
- UI button improvements.
