+++ ConcordiumWallet/Views/WalletConnect/ApproveScreen/WalletConnectActionRequestView.swift
@@ -9,23 +9,14 @@
 import SwiftUI
 import Web3Wallet
 struct WalletConnectActionRequestView: View {
-    let dappName: String
     let amount: GTU
     let balanceAtDisposal: GTU
     let contractAddress: ContractAddress
     let transactionType: String
     let params: ContractUpdateParameterRepresentation
-    let didAccept: () -> Void
-    let didReject: () -> Void
     let request: Request
     var body: some View {
         VStack {
-            Text("Transaction Approval")
-                .bold()
-                .font(.system(size: 20))
-            
-            Text("\(dappName) requests your signature on the following transaction: ")
-                .padding()
             HStack {
                 Text("Account Balance:")
                 Text("\(balanceAtDisposal.displayValueWithGStroke())")
@@ -64,39 +55,7 @@ struct WalletConnectActionRequestView: View {
                     .stroke(.gray, lineWidth: 1)
             )
             .padding()
-
-            Spacer()
-
-            HStack(spacing: 16) {
-                Button(action: {
-                    didReject()
-                }, label: {
-                    Text("Reject")
-                        .foregroundColor(Pallette.error)
-                        .padding()
-                        .frame(maxWidth: .infinity)
-                        .background(
-                            RoundedRectangle(
-                                cornerRadius: 10,
-                                style: .continuous
-                            )
-                            .stroke(Pallette.error, lineWidth: 2)
-                        )
-                })
-
-                Button(action: {
-                    didAccept()
-                }, label: {
-                    Text("Accept")
-                        .foregroundColor(.white)
-                        .padding()
-                        .frame(maxWidth: .infinity)
-                })
-                .background(Pallette.primary)
-                .cornerRadius(10)
-            }
         }
-        .padding()
     }
 
     func buildTransactionItem(title: String, value: some View) -> some View {
